Added: New testing files Changed to conditional return from post_details?
